Implement basic error handling
Try and catches were thrown everywhere for this. In the future a developer option
should be added to include print out the errors. Also be sure to look at line 30 in this version.
I included there a note about how I didn't see good documentation about the exception ParseError
from the xml.etree package in the python libraries. This should be looked into as well.