Fix breaking code from error handling
This was a general code cleanup effort. ParseError was supposed to be in the ET
package, I just listed it as a ParseError. The same thing happened with URLError
and finally I fixed the part that involved inputing an APIKEY if it did not exist.
An unused import was also removed, hurray!

Credit as usual to TheLocust3, for his APIKEY idea about inputing it in the program, he gave me a pull request, but
due to several reasons I could not merge it.