Fix python3 conversion and overhaul code
There are a lot of updates in this commit, sadly without each change,
there could be no working commit, it had to happen all at once. In this
commit the etree package is no longer used and instead the lxml etree package
is used, this is because it's a better library and python3 was having difficulty
with the normal etree package.
The second major change is that, since python3 needs it, urllib was changed to
the package urllib.request.
The third major change to keep up with the python3 compatability was the swapping of
OptionParser for argparser, this change was small but is worth a note.

In summary the big changes were:
* use of lxml
* use of urllib.request
* use of argparser

Among that are fixes for things that were breaking in the previous commit.