Update README to have better install instructions
The major difference is that I now set up a BitBucket repository of this, and now it
isn't centralized on GitHub. I also fixed some spacing things that were annoying me (I
removed the spaces underneath the headers because why not?). I also updated the install
instructions to suggest using pip to install the project, as that is the point of using a
setup.py and package structure. For easy installs!